Create code
This code generates 10 random integers between 1 and 100 (inclusive) using the C++11 random library. The random_device class is used to obtain a seed for the random number generator, and the mt19937 class is a standard random number generator based on the Mersenne Twister algorithm. The uniform_int_distribution class is used to generate random integers within a specified range.

The program can be made more flexible by declaring the number of random numbers to be generated and the range of numbers as variables and taking the inputs from the user.